+++ src/app/actions.ts
@@ -1,32 +1,15 @@
 'use server'
 import { apiUrl } from '@/config'
 import { CreateAttachmentRequest } from '@/types/dto/attachments.dto'
-import { CreateTaskRequest, UpdateTaskRequest, TaskResponse } from '@/types/dto/tasks.dto'
+import { CreateTaskRequest, UpdateTaskRequest } from '@/types/dto/tasks.dto'
 import { CreateViewSettingsDTO } from '@/types/dto/viewSettings.dto'
-import { View } from '@/types/interfaces'
-import { revalidateTag } from 'next/cache'
 
 export const handleCreate = async (token: string, payload: CreateTaskRequest) => {
   const response = await fetch(`${apiUrl}/api/tasks?token=${token}`, {
     method: 'POST',
     body: JSON.stringify(payload),
   })
-  revalidateTag('getTasks')
   return await response.json()
-}
-
-/**
- * @deprecated
- * Use the new update task function instead. This will be completely removed in the upcoming PRs.
- */
-export const updateWorkflowStateIdOfTask = async (token: string, taskId: string, targetWorkflowStateId: string) => {
-  await fetch(`${apiUrl}/api/tasks/${taskId}?token=${token}`, {
-    method: 'PATCH',
-    body: JSON.stringify({
-      workflowStateId: targetWorkflowStateId,
-    }),
-  })
-  revalidateTag('getAllTasks')
 }
 
 export const updateTask = async ({
@@ -49,7 +32,6 @@ export const updateTask = async ({
       dueDate: payload.dueDate,
     }),
   })
-  revalidateTag('getAllTasks')
 }
 
 export const updateViewModeSettings = async (token: string, payload: CreateViewSettingsDTO) => {
@@ -57,9 +39,6 @@ export const updateViewModeSettings = async (token: string, payload: CreateViewS
     method: 'PATCH',
     body: JSON.stringify(payload),
   })
-  revalidateTag('getViewSettings')
-  revalidateTag('getAllTasks')
-  revalidateTag('getAllWorkflowStates')
 }
 
 export const getSignedUrlUpload = async (token: string, fileName: string) => {
@@ -73,5 +52,4 @@ export const createMultipleAttachments = async (token: string, attachments: Crea
     method: 'POST',
     body: JSON.stringify(attachments),
   })
-  revalidateTag('getAllTasks')
 }
